pyproject.toml
[tool.poetry]
name = "django-svelte-sample-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
whitenoise = "^5.3.0"
djangorestframework = "^3.12.4"
pytz = "^2021.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
